chore: move generate sources to a new class
