+++ sdk/Dagger.SDK.Mod.SourceGenerator/SourceGenerator.cs
@@ -24,8 +24,10 @@ public class SourceGenerator : IIncrementalGenerator
     {
         context.RegisterPostInitializationOutput(postInitializationContext =>
         {
-            postInitializationContext.AddSource("Dagger.SDK.Mod_Attributes.g.cs", ModuleAttributesSource());
-            postInitializationContext.AddSource("Dagger.SDK.Mod_Interfaces.g.cs", ModuleInterfacesSource());
+            postInitializationContext.AddSource("Dagger.SDK.Mod_Attributes.g.cs",
+                GenerateSources.ModuleAttributesSource());
+            postInitializationContext.AddSource("Dagger.SDK.Mod_Interfaces.g.cs",
+                GenerateSources.ModuleInterfacesSource());
         });
         var objectClasses = context.SyntaxProvider.ForAttributeWithMetadataName(
             fullyQualifiedMetadataName: ObjectAttribute,
@@ -34,49 +36,6 @@ public class SourceGenerator : IIncrementalGenerator
         );
 
         context.RegisterSourceOutput(objectClasses, GenerateIDagSetter);
-    }
-
-    private static SourceText ModuleInterfacesSource()
-    {
-        const string sourceText = """
-                                  namespace Dagger.SDK.Mod;
-
-                                  /// <summary> 
-                                  /// An interface for module runtime to inject Dagger client instance to the 
-                                  /// object class.
-                                  /// </summary>
-                                  public interface IDagSetter
-                                  {
-                                      /// <summary>
-                                      /// Set Dagger client instance.
-                                      /// </summary>
-                                      /// <param name="dag">The Dagger client instance.</param>
-                                      void SetDag(Query dag);
-                                  }
-                                  """;
-        return SourceText.From(sourceText, Encoding.UTF8);
-    }
-
-    private static SourceText ModuleAttributesSource()
-    {
-        const string sourceText = """
-                                  using System;
-
-                                  namespace Dagger.SDK.Mod;
-
-                                  /// <summary>
-                                  /// Expose the class as a Dagger.ObjectTypeDef.
-                                  /// </summary>
-                                  [AttributeUsage(AttributeTargets.Class)]
-                                  public sealed class ObjectAttribute : Attribute;
-
-                                  /// <summary>
-                                  /// Expose the class as a Dagger.Function.
-                                  /// </summary>
-                                  [AttributeUsage(AttributeTargets.Method)]
-                                  public sealed class FunctionAttribute : Attribute;
-                                  """;
-        return SourceText.From(sourceText, Encoding.UTF8);
     }
 
     private static (ClassDeclarationSyntax classDef, INamedTypeSymbol classSymbol) ExtractTarget(
